+++ www/src/templates/template-contributor-page.js
@@ -30,17 +30,13 @@ class ContributorPageTemplate extends React.Component {
           {contributor.avatar && (
             <meta
               property="og:image"
-              content={`https://gatsbyjs.org${
-                contributor.avatar.childImageSharp.fixed.src
-              }`}
+              content={`https://gatsbyjs.org${contributor.avatar.childImageSharp.fixed.src}`}
             />
           )}
           {contributor.avatar && (
             <meta
               name="twitter:image"
-              content={`https://gatsbyjs.org${
-                contributor.avatar.childImageSharp.fixed.src
-              }`}
+              content={`https://gatsbyjs.org${contributor.avatar.childImageSharp.fixed.src}`}
             />
           )}
         </Helmet>