+++ packages/gatsby-plugin-emotion/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-emotion@4.1.4...gatsby-plugin-emotion@4.1.5) (2019-09-01)
+
+### Bug Fixes
+
+- update minor updates in packages except react, babel and eslint ([#17254](https://github.com/gatsbyjs/gatsby/issues/17254)) ([252d867](https://github.com/gatsbyjs/gatsby/commit/252d867))
+
+## [4.1.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-emotion@4.1.3...gatsby-plugin-emotion@4.1.4) (2019-08-23)
+
+**Note:** Version bump only for package gatsby-plugin-emotion
+
+## [4.1.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-emotion@4.1.2...gatsby-plugin-emotion@4.1.3) (2019-08-20)
+
+### Bug Fixes
+
+- update dependency [@emotion](https://github.com/emotion)/babel-preset-css-prop to ^10.0.14 ([#16795](https://github.com/gatsbyjs/gatsby/issues/16795)) ([82dac42](https://github.com/gatsbyjs/gatsby/commit/82dac42))
+
 ## [4.1.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-emotion@4.1.1...gatsby-plugin-emotion@4.1.2) (2019-07-12)
 
 ### Bug Fixes