+++ packages/gatsby-plugin-typography/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.5...gatsby-plugin-typography@2.3.6) (2019-09-01)
+
+### Bug Fixes
+
+- update minor updates in packages except react, babel and eslint ([#17254](https://github.com/gatsbyjs/gatsby/issues/17254)) ([252d867](https://github.com/gatsbyjs/gatsby/commit/252d867))
+
+## [2.3.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.4...gatsby-plugin-typography@2.3.5) (2019-08-23)
+
+**Note:** Version bump only for package gatsby-plugin-typography
+
+## [2.3.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.3...gatsby-plugin-typography@2.3.4) (2019-08-23)
+
+**Note:** Version bump only for package gatsby-plugin-typography
+
+## [2.3.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.2...gatsby-plugin-typography@2.3.3) (2019-08-20)
+
+**Note:** Version bump only for package gatsby-plugin-typography
+
 ## [2.3.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.1...gatsby-plugin-typography@2.3.2) (2019-07-12)
 
 ### Bug Fixes